backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "caresync-backend"
version = "0.1.0"
description = "Modular FastAPI backend for CareSync NHS workflow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy[asyncio]>=2.0.40",
  "aiosqlite>=0.21.0",
  "asyncpg>=0.30.0",
  "psycopg[binary]>=3.2.9",
  "alembic>=1.16.0",
  "pydantic-settings>=2.10.0",
  "python-jose[cryptography]>=3.5.0",
  "passlib>=1.7.4",
  "python-multipart>=0.0.20",
  "email-validator>=2.2.0",
  "openai>=1.80.0",
  "python-dateutil>=2.9.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.28.0",
  "ruff>=0.12.0",
  "mypy>=1.17.0"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "path*"]
